64 bit fixes

+++ src/idlib/Heap.cpp
@@ -42,12 +42,14 @@ If you have questions concerning this license or the applicable additional terms
 Mem_Alloc16
 ==================
 */
-void * Mem_Alloc16( const int size, const memTag_t tag ) {
-	if ( !size ) {
+void* Mem_Alloc16(const size_t size, const memTag_t tag)
+{
+	if (!size)
+	{
 		return NULL;
 	}
-	const int paddedSize = ( size + 15 ) & ~15;
-	return _aligned_malloc( paddedSize, 16 );
+	const size_t paddedSize = (size + 15) & ~15;
+	return _aligned_malloc(paddedSize, 16);
 }
 
 /*
@@ -55,11 +57,13 @@ void * Mem_Alloc16( const int size, const memTag_t tag ) {
 Mem_Free16
 ==================
 */
-void Mem_Free16( void *ptr ) {
-	if ( ptr == NULL ) {
+void Mem_Free16(void* ptr)
+{
+	if (ptr == NULL)
+	{
 		return;
 	}
-	_aligned_free( ptr );
+	_aligned_free(ptr);
 }
 
 /*
@@ -67,9 +71,10 @@ void Mem_Free16( void *ptr ) {
 Mem_ClearedAlloc
 ==================
 */
-void * Mem_ClearedAlloc( const int size, const memTag_t tag ) {
-	void * mem = Mem_Alloc( size, tag );
-	SIMDProcessor->Memset( mem, 0, size );
+void* Mem_ClearedAlloc(const size_t size, const memTag_t tag)
+{
+	void* mem = Mem_Alloc(size, tag);
+	SIMDProcessor->Memset(mem, 0, size);
 	return mem;
 }
 
@@ -78,9 +83,10 @@ void * Mem_ClearedAlloc( const int size, const memTag_t tag ) {
 Mem_CopyString
 ==================
 */
-char *Mem_CopyString( const char *in ) {
-	char * out = (char *)Mem_Alloc( strlen(in) + 1, TAG_STRING );
-	strcpy( out, in );
+char* Mem_CopyString(const char* in)
+{
+	char* out = (char*)Mem_Alloc(strlen(in) + 1, TAG_STRING);
+	strcpy(out, in);
 	return out;
 }
